Solve LAP externally by python script
Write distance matrix to file 'file.dat' for solving the LAP by the JVC algorithm in the python script 'JVC.py'. 
Read the cost/sum of distances ('cost.csv') and the assignment of the atoms ('map.csv') and assign them to the variables dist_map and map, respectively.
Remove these help files afterwards.